fix(automations): keep loading accurate across concurrent list requests
